refactor : Member Storage seperate AllMember Class
