+++ src/main/java/toolc/yourlist/member/MemberAuthProvider.java
@@ -2,12 +2,11 @@ package toolc.yourlist.member;
 
 import io.vavr.control.Either;
 import lombok.RequiredArgsConstructor;
+import toolc.yourlist.member.domain.AllMember;
 import toolc.yourlist.member.domain.loginId.LoginId;
 import toolc.yourlist.member.domain.password.Password;
 
 import java.time.Period;
-import java.util.HashMap;
-import java.util.Map;
 
 import static io.vavr.control.Either.left;
 import static io.vavr.control.Either.right;
@@ -17,38 +16,27 @@ public class MemberAuthProvider {
 
   private final TokenProvider tokenProvider;
   private final TokenReader tokenReader;
-
-  Map<Long, String> memberStorage = new HashMap<>();
-  Long id = 1L;
+  private final AllMember allMember;
 
   Either<String, String> registerMember(LoginId loginId, Password password) {
-    if (memberStorage.containsValue(loginId.raw())) {
+    if (allMember.isNotExistByLoginId(loginId)) {
       return left("Already register Member");
     } else {
-      memberStorage.put(id++, loginId.raw());
-      return right(loginId.raw());
+      allMember.registerMember(loginId, password);
+      return right("Success register new Member");
     }
   }
 
 
   Either<String, Token> login(LoginId loginId, Password password, boolean isPC) {
-    Long id = findIdByLoginId(loginId);
+    Long id = allMember.findIdByLoginId(loginId);
     Period refreshTokenExpiration = isPC ? Period.ofDays(7) : Period.ofDays(90);
     return right(tokenProvider.makeToken(id, refreshTokenExpiration));
   }
 
-  private Long findIdByLoginId(LoginId loginId) {
-    return memberStorage.entrySet()
-      .stream()
-      .filter(enyty -> enyty.getValue().equals(loginId.raw()))
-      .map(Map.Entry::getKey)
-      .findFirst()
-      .get();
-  }
-
   Either<String, Token> reissueToken(String accessToken, String refreshToken, boolean isPC) {
     Long id = tokenReader.getId(accessToken);
-    if (!memberStorage.containsKey(id)) {
+    if (allMember.isNotExistById(id)) {
       throw new IllegalArgumentException();
     }
