+++ src/main/java/toolc/yourlist/member/domain/AllMember.java
@@ -1,10 +1,12 @@
-package toolc.yourlist.member;
+package toolc.yourlist.member.domain;
 
 import toolc.yourlist.member.domain.loginId.LoginId;
 import toolc.yourlist.member.domain.password.Password;
 
 public interface AllMember {
-  Member isExistMember(String loginId);
+  boolean isNotExistByLoginId(LoginId loginId);
+
+  boolean isNotExistById(Long id);
 
   void registerMember(LoginId loginId, Password password);
